Fix some specs
The collection factory was failing because we overrode the
basic_metadata.yaml which it also uses so we need to add all the basic
metadata into the collection_resource.yaml.  There was a log that was
failing because it was not being explicit in the root of the application
though I'm unsure if it is needed.  The major question is if we need to
run hyku_specs or not.