Add the HykuKnapsack::NestedWorks to all concerns
The `HykuKnapsack::NestedWorks` redefines the `valid_child_concerns`
method to evalute the Hyrax.config.curation_concerns at runtime.  In
Hyrax, the `valid_child_concerns` is defined via a `class_attribute`
which seems to get set at time of initializing which at the time it gets
set, the Knapsack additional curation concerns are not added.