✅ Fix flaky spec with Advanced Search
There was a flaky spec that ocurred when the CdlsController spec ran. As
far as I can tell, it was happening when a CDL factory was being
created.  I didn't chase this one down too far but the fix was making a
watered down SolrDocument instead of creating a CDL and then turning
that into a SolrDcoument.  This saves a trip to the database and makes
the spec a little faster as well.